// src/app/pages/works-cited/types/data-source-type.enum.ts
export enum DataSourceType {
  MDH = 'Memphis Data Hub',
  CENSUS = 'US Census',
  SCS = 'Shelby County Schools',
  TNEDU = 'Tennessee Department of Education',
  ASPE = 'Office of the Assistant Secretary for Planning and Evaluation',
  UWMS = 'United Way of the Mid-South',
  CR = 'Census Reporter'
}